Linting, formatting, updated dependencies and increased coverage slightly

+++ refined/src/main/scala/itinere/refined/package.scala
@@ -3,13 +3,12 @@ package itinere
 import eu.timepit.refined.api.{RefType, Refined, Validate}
 import eu.timepit.refined.numeric.Positive
 
-
 package object refined {
 
   implicit class RichIntJson(val Json: Json[Int]) {
     private def withRefined[P](bound: Range)(implicit V: Validate[Int, P], R: RefType[Refined]): Json[Int Refined P] =
       new Json[Refined[Int, P]] {
-        override def apply[F[_] : JsonAlgebra]: F[Refined[Int, P]] =
+        override def apply[F[_]: JsonAlgebra]: F[Refined[Int, P]] =
           JsonAlgebra[F].pmap(JsonAlgebra[F].int(bound))(p => Attempt.fromEither(R.refine(p)))(R.unwrap)
       }
 
@@ -20,11 +19,11 @@ package object refined {
   implicit class RichLongJson(val Json: Json[Long]) {
     private def withRefined[P](bound: Range)(implicit V: Validate[Long, P], R: RefType[Refined]): Json[Long Refined P] =
       new Json[Refined[Long, P]] {
-        override def apply[F[_] : JsonAlgebra]: F[Refined[Long, P]] =
+        override def apply[F[_]: JsonAlgebra]: F[Refined[Long, P]] =
           JsonAlgebra[F].pmap(JsonAlgebra[F].long(bound))(p => Attempt.fromEither(R.refine(p)))(R.unwrap)
       }
 
-    def positive[N <: Int](implicit V: Validate.Plain[Int, Positive]): Json[Long Refined Positive] =
+    def positive[N <: Int](implicit V: Validate.Plain[Long, Positive]): Json[Long Refined Positive] =
       withRefined(Range(Bound(1l, true), Bound(Long.MaxValue, true)))
   }
 }
